added one more opcode test, made tests pass

+++ src/main/scala/io/iohk/ethereum/vm/Program.scala
@@ -15,5 +15,5 @@ case class Program(code: ByteString) {
     }
 
   def getBytes(from: Int, size: Int): ByteString =
-    code.slice(from, from + size).padTo(size, 0)
+    code.slice(from, from + size).padTo(size, 0.toByte)
 }
